refactor: BookRepository 인터페이스 domain layer 로 이동
